1. Added tests for 2 alarms at the same time 2. Huge refactoring, including reworked dismissed alarms (to store the alarm id instead of alarm time)
Fixed: Currently there is an assumption: there is at most one alarm on a particular alarm time. With introduction of one-time alarms this is not true anymore. There may be large implications for entire app, not only for the dismissed alarms.

Fixed: Improve architecture of one-time alarms (when it is changed while snoozed) - store the one-time alarm ID so we can return the proper object

Fixed: Test: 1. set alarm to a time, 2. dismiss this alarm, 3. set alarm to the same time. Is it really working?